requirements.txt
Django
requests>=2.21
djangorestframework
pwgen>0.7
gunicorn
xmltodict
dj_rest_auth
